backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
playwright>=1.40.0
google-generativeai>=0.5.0
anthropic>=0.25.0
python-multipart>=0.0.6
sse-starlette>=1.6.0
aiofiles>=23.0.0
Pillow>=10.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
